Adding button requirements
Adding requirements to trigger button or switch activations. So if we want to activate switch A, we can add a requirement that switch B is already activated. This could be useful in cases like requiring that we can't press a button to mute a microphone whilst the push to talk button is also held.