Removed Authentication Tokens
These where overkill on securtity anyway especially considering the box will realistically only be used on a private network. Now the Host PC will be connecting to the Button Box, not the other way around, therefor the host PC is always in control of the box it is connected to.

In either case, the user needs to start a flask server anyway so they are well aware of the operations being carried out on their machine.