Made button cancellations more reliable
Now buttons respect the order in which they are pressed, based on the latest event. If a previous button is waiting to reset the default message, the current button event will override this with its own message.

Still a bit glitchy - need to find out why sometimes fast button presses crash the button box