Hopefully fixing display issues
Sometimes when flicking buttons abnormally fast the display gets mangled and displays text backwards. I'm not sure why but I think it is because of the async nature of Flask API's, so when hitting buttons fast, the callback API quickly sends loads of display message requests which inadvertantly break the display by botching the cursor position in an unfixable way?

Now clearing the display before each message, I will also make a change to ensure that we do not send a change message request whilst another one has not been responded to and confirmed as completed - if this does not fix the problem