Hopefully fixing mangled display
I am trying to fix an issue whereby when pressing buttons abnormally fast, the display sometimes gets into an unrecoverable state whereby the text displays backwards. I theorise that this is due to the async nature of the flask API that updates the display, sometimes causing overlapping requests and as such potentially breaking the cursor position.

Previously I tried to clear the display and reset the cursor position each display update, but this takes too long. Now I've introduced a variable to ensure the display is only ever being updated by one request. If the service is called whilst the display is being updated, it will wait retry 5 times within the space of a second and if the display is still being updated, the request will be cancelled to ensure the display cursor position is not broken.

This may mean that some requests jump the queue if there is unfortunate timing, but realistically the user should never be pressing buttons that fast in the first place.