Speeding up display update times
The display is quite slow to update when constantly sending update messages. Now we only update the lines that have changed.

In the future I could make it so that it only updates the individual line segments that have changed through manipulating the screen cursor, but right now that would be overkill.