Fixing MS Flight Service Bug
I have a suspicion the plane settings are staying the same even when you change planes, currently this can be fixed with a restart of the code but I want something more robust. Now every time you change configuration the flight sim service resets, so if you press a flight sim button again when switching back to the configuration, it will reset for the current plane you are in.