Starting to add a system to handle integrations
I am starting to test out a modular approach for integrations. So in an ideal world, to create an integration a user will only have to touch 3 places - the folder the actual integration is contained in to implement it, the templates folder to create any custom web-pages the integration requires and finally the integration_factory to implement the integration.

I do not want to end up in a situation with tons of interweaving code for integrations, whereby adding/removing/editing an integration will cause side effects. I want the concept of an integration to be a completely individual aspect of the system - if that makes any sense at all.