Added optional name for buttons
It became apparent especially with the keybind integration that buttons may also need a description. So if a keybind integration action is named "CTRL+X", we might want to describe what this does in the context of a configuration. So in the KSP config it might abort a mission, but in a different games configuration the keybind might do something different. So the ability to name a button has been added.