Adding another initialisation step for integrations.
The base initialise() method ensures that the admin panel database has an up to date state of each integration for display and UI purposes. On the other hand, the initialise_service() method is unique to each service and will provide that service with an opportunity to connect to any external API's they need to do etc, at startup of the admin panel.