✅ Adding Mode Selection Integration
This configuration introduces new Integration Actions based on the configurations saved in the system. It allows a user to map a button to switch the box to a different configuration.

It also showcases the modularity of this admin panel for the button box. To add a whole new feature, all that was needed was copying a standard class and adding custom functionality, and then just adding that class to a list in the integration factory.