Have to rejig how the Keyboard and future integrations import their service
Previously the KeyboardService was initialised in the IntegrationFactory, but this won't work because there needs to be access to this service from the Keyboard.controllers.py file so that we can route API calls tot he service.

I'm doing the same as the core module now, so the controllers.py is considered the source of truth. The controllers.py file for the Keyboard (and other integrations requiring a UI in the admin panel) will now be the only place to instanciate the KeyboardService.

For any integrations with a UI configuration element, the controllers.py file of that integration should instantiate the service. And then we will import that instanciated service and add it to the integrations list.